feat: shifted bsky classes in agent
