+++ packages/client/src/bsky/index.ts
@@ -1,2 +1,4 @@
-export * from './bsky';
+export * from './actor';
 export * from './feed';
+export * from './list';
+export * from './post';
