+++ packages/client/src/user/index.ts
@@ -1,5 +1,5 @@
 import type { AppBskyFeedGetTimeline } from '@tsky/lexicons';
-import { Actor } from '~/bsky/actor';
+import { Actor } from '~/bsky';
 import type { RPCOptions } from '~/types';
 import { Paginator } from '~/utils';
 import { Mute } from './mute';
